// apps/webapp/src/wui/_autogen/api.ts
export interface UserData {
	id: string;
	username: string;
	name?: string;
	avatarUrl: string;
	profileUrl: string;
}
// ...
export type ModDataCategory = (typeof ModDataCategory)[keyof typeof ModDataCategory];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ModDataCategory = {
	CAMPAIGN: "CAMPAIGN",
	DEVICE_PROFILES: "DEVICE_PROFILES",
	MOD: "MOD",
	MISSION: "MISSION",
	SKIN: "SKIN",
	SOUND: "SOUND",
	TERRAIN: "TERRAIN",
	UTILITY: "UTILITY",
	OTHER: "OTHER",
} as const;

export type ModDataVisibility = (typeof ModDataVisibility)[keyof typeof ModDataVisibility];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ModDataVisibility = {
	PUBLIC: "PUBLIC",
	PRIVATE: "PRIVATE",
	UNLISTED: "UNLISTED",
} as const;

export type ModDataLatestReleaseId = string | null;

/**
 * Data representation of a mod.
 */
export interface ModData {
	id: string;
	name: string;
	category: ModDataCategory;
	description: string;
	content: string;
	tags: string[];
	dependencies: string[];
	screenshots: string[];
	thumbnail: string;
	visibility: ModDataVisibility;
	/** @minItems 1 */
	maintainers: string[];
	latestReleaseId?: ModDataLatestReleaseId;
	downloadsCount: number;
}
// ...
export type ModReleaseAssetDataUrlsItem = {
	/** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
	id: string;
	url: string;
};

/**
 * Data representation of a mod release asset.
 */
export interface ModReleaseAssetData {
	/** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
	id: string;
	/** @minLength 1 */
	name: string;
	urls: ModReleaseAssetDataUrlsItem[];
	isArchive: boolean;
}

export type ModReleaseSymbolicLinkDataDestRoot =
	(typeof ModReleaseSymbolicLinkDataDestRoot)[keyof typeof ModReleaseSymbolicLinkDataDestRoot];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ModReleaseSymbolicLinkDataDestRoot = {
	DCS_WORKING_DIR: "DCS_WORKING_DIR",
	DCS_INSTALL_DIR: "DCS_INSTALL_DIR",
} as const;

/**
 * Data representation of a symbolic link configuration.
 */
export interface ModReleaseSymbolicLinkData {
	/** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
	id: string;
	/** @minLength 1 */
	name: string;
	/** @minLength 1 */
	src: string;
	/** @minLength 1 */
	dest: string;
	destRoot: ModReleaseSymbolicLinkDataDestRoot;
}

export type ModReleaseMissionScriptDataRoot =
	(typeof ModReleaseMissionScriptDataRoot)[keyof typeof ModReleaseMissionScriptDataRoot];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ModReleaseMissionScriptDataRoot = {
	DCS_WORKING_DIR: "DCS_WORKING_DIR",
	DCS_INSTALL_DIR: "DCS_INSTALL_DIR",
} as const;

export type ModReleaseMissionScriptDataRunOn =
	(typeof ModReleaseMissionScriptDataRunOn)[keyof typeof ModReleaseMissionScriptDataRunOn];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ModReleaseMissionScriptDataRunOn = {
	MISSION_START_BEFORE_SANITIZE: "MISSION_START_BEFORE_SANITIZE",
	MISSION_START_AFTER_SANITIZE: "MISSION_START_AFTER_SANITIZE",
} as const;

/**
 * Data representation of a mission script configuration.
 */
export interface ModReleaseMissionScriptData {
	/** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
	id: string;
	/** @minLength 1 */
	name: string;
	/** @minLength 1 */
	purpose: string;
	/** @minLength 1 */
	path: string;
	root: ModReleaseMissionScriptDataRoot;
	runOn: ModReleaseMissionScriptDataRunOn;
}

export type ModReleaseDataVisibility = (typeof ModReleaseDataVisibility)[keyof typeof ModReleaseDataVisibility];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ModReleaseDataVisibility = {
	PUBLIC: "PUBLIC",
	PRIVATE: "PRIVATE",
	UNLISTED: "UNLISTED",
} as const;

/**
 * Data representation of a mod release.
 */
export interface ModReleaseData {
	id: string;
	modId: string;
	version: string;
	versionHash: string;
	changelog: string;
	assets: ModReleaseAssetData[];
	symbolicLinks: ModReleaseSymbolicLinkData[];
	missionScripts: ModReleaseMissionScriptData[];
	visibility: ModReleaseDataVisibility;
	downloadsCount?: number;
	createdAt?: string;
	updatedAt?: string;
}